change ConnectedClientID type to uint32_t

+++ networking/ntshengn_networking_common.h
@@ -38,7 +38,7 @@
 
 namespace NtshEngn {
 
-	typedef uint64_t ConnectedClientID;
+	typedef uint32_t ConnectedClientID;
 
 	enum class NetworkType {
 		UDP,
